Module rtlsdr::dsp::block

At the moment the entire DSP code base is built as blocks, which form pipelines. These pipelines can be branched and converged, and form a graph. The entire namespace is designed with as much performance as possible while still being safe and ergonomic in spirit of the Rust language.

Modules

chunker
This module contains blocks that will take input and either break or collect it into specific sized chunks. This is useful when you are unable to provide a certain sized chunk of input to a block, but the block requires that specific size.

demod
Demodulate signals.

fft
This module holds anything related to a FFT, including a DFT.

filter
Filter signals.

mixer
Signal generator or mixer.

source
Provides signal data.

transform
This module transforms input chunks.

Enums

BlockOutput
This is what all sources and adapters return. It supports the ability for the output to be valid or invalid, and also provides the ability to signal that more input data is needed.

SourceAmount
This is used by a source so that you can specify exactly how much or how little data you desire. It provides support for also requesting what the source block determines if an optimal size for performance, and it provides support for finite sources so that you can read all data from the source.

Traits

Adapter
This is a block that produces output for input.

Sink
This is a block that consumes input with no direct output.

Source
This is a block that produces output with no direct input.