Enum rtlsdr::dsp::block::BlockOutput

pub enum BlockOutput<O> {
    ErrorInputSizeInvalid(String),
    NeedMoreInput(usize),
    Ready(Vec<O>),
}

This is what all sources and adapters return. It supports the ability for the output to be valid or invalid, and also provides the ability to signal that more input data is needed.

Variants

ErrorInputSizeInvalid

The amount of data provided is not supported, because it was it was either too many or too few, and it has been discarded. The string will hopefully contain a useful diagnostic message.

NeedMoreInput

The data was processed, but more if needed to produce some output.

Ready

Methods

impl<O> BlockOutput<O>

fn unwrap(self) -> Vec<O>