Struct rtlsdr::dsp::block::source::rtlsdr::usb::USB

pub struct USB {
    // some fields omitted
}

Methods

impl USB

fn asyncreadchunk(&self) -> Vec<SAMPLE>

Read a single chunk of unknown size, ONLY if async was request when by the RTLSDR::new being called with async set to true. If not you must either manual setup async reading or call the syncread method.

This is going to likely be what the offloader thread read from a single USB event. It is generally going to be controlled from somewhere else.

This provides an efficient way to process the offload queue.

fn asyncread(&self, maxcnt: usize) -> Vec<SAMPLE>

Read a specified maximum number of samples, ONLY if async was requested by the RTLSDR::new being called with async set to true. If not you must either manual setup async reading or call the syncread method.

This may be a little inefficient, since it might have to chop up some buffers to pull out this exact maximum number of bytes if avaliable. If you need to be efficient then consider using asyncreadchunk.

fn syncread(&self, maxcnt: isize) -> Vec<SAMPLE>

Read a specified maximum number of samples. WARNING: If used with async mode then the behavior is unknown currently.

TODO: Make function to read into buffer, instead of creating a new buffer for each invocation.

fn setcenterfreq(&self, freq: u32)

Set the center frequency of the device.

fn getsamplerate(&self) -> u32

Get the current sample rate of the device.

fn setsamplerate(&self, freq: u32)

Sets the sampling frequency.

fn setmanualtuner(&self, on: bool)

Will enable or disable manual tuning.

fn new(index: u32, asyncmemsize: usize) -> Option<USB>

Create a new controling instance of the device.

fn isasync(&self) -> bool